+++ keconnection/keconnection-heatmeter/src/main/java/org/complitex/keconnection/heatmeter/entity/HeatmeterConnection.java
@@ -9,7 +9,7 @@ import static org.complitex.keconnection.heatmeter.entity.HeatmeterPeriodType.CO
 public class HeatmeterConnection extends HeatmeterPeriod {
     private Long buildingId;
     private Long organizationId;
-    private Integer code;
+    private Long code;
     private String organizationCode;
 
     public HeatmeterConnection() {
@@ -32,11 +32,11 @@ public class HeatmeterConnection extends HeatmeterPeriod {
         this.organizationId = organizationId;
     }
 
-    public Integer getCode() {
+    public Long getCode() {
         return code;
     }
 
-    public void setCode(Integer code) {
+    public void setCode(Long code) {
         this.code = code;
     }
